fix: 명세서에 맞게 변수 이름 수정

+++ src/main/java/com/followMe/delivery_server/delivery/domain/DeliveryManagerCandidate.java
@@ -2,4 +2,4 @@ package com.followMe.delivery_server.delivery.domain;
 
 import java.util.UUID;
 
-public record DeliveryManagerCandidate(UUID id, String name, int sequence, int assignedCount) {}
+public record DeliveryManagerCandidate(UUID id, String name, long sequence, int assignedCount) {}
